refactor: move SetInputsOrFallback function to inputs package and update references
